chore(webgl): Join texture utility files
