+++ modules/webgl/src/adapter/helpers/format-utils.ts
