+++ modules/webgl/src/deprecated/clear.ts
